clippy autofixes

+++ packed_struct_tests/tests/primitive_enum_types.rs
@@ -53,33 +53,33 @@ pub enum EnumI64 {
 
 #[test]
 fn prim_ty() {
-    assert_eq!(0 as u8, EnumU8::VariantMin.to_primitive());
-    assert_eq!(255 as u8, EnumU8::VariantMax.to_primitive());
+    assert_eq!(0_u8, EnumU8::VariantMin.to_primitive());
+    assert_eq!(255_u8, EnumU8::VariantMax.to_primitive());
 
-    assert_eq!(0 as u16, EnumU16::VariantMin.to_primitive());
-    assert_eq!(65535 as u16, EnumU16::VariantMax.to_primitive());
+    assert_eq!(0_u16, EnumU16::VariantMin.to_primitive());
+    assert_eq!(65535_u16, EnumU16::VariantMax.to_primitive());
 
-    assert_eq!(0 as u32, EnumU32::VariantMin.to_primitive());
-    assert_eq!(4294967295 as u32, EnumU32::VariantMax.to_primitive());
+    assert_eq!(0_u32, EnumU32::VariantMin.to_primitive());
+    assert_eq!(4294967295_u32, EnumU32::VariantMax.to_primitive());
 
     #[cfg(target_pointer_width = "64")]
     {
-        assert_eq!(0 as u64, EnumU64::VariantMin.to_primitive());
-        assert_eq!(1844674407370955165 as u64, EnumU64::VariantMax.to_primitive());
+        assert_eq!(0_u64, EnumU64::VariantMin.to_primitive());
+        assert_eq!(1844674407370955165_u64, EnumU64::VariantMax.to_primitive());
     }
 
-    assert_eq!(-128 as i8, EnumI8::VariantMin.to_primitive());
-    assert_eq!(127 as i8, EnumI8::VariantMax.to_primitive());
+    assert_eq!(-128_i8, EnumI8::VariantMin.to_primitive());
+    assert_eq!(127_i8, EnumI8::VariantMax.to_primitive());
 
-    assert_eq!(-32768 as i16, EnumI16::VariantMin.to_primitive());
-    assert_eq!(32767 as i16, EnumI16::VariantMax.to_primitive());    
+    assert_eq!(-32768_i16, EnumI16::VariantMin.to_primitive());
+    assert_eq!(32767_i16, EnumI16::VariantMax.to_primitive());    
 
-    assert_eq!(-2147483648 as i32, EnumI32::VariantMin.to_primitive());
-    assert_eq!(2147483647 as i32, EnumI32::VariantMax.to_primitive());
+    assert_eq!(-2147483648_i32, EnumI32::VariantMin.to_primitive());
+    assert_eq!(2147483647_i32, EnumI32::VariantMax.to_primitive());
 
     #[cfg(target_pointer_width = "64")]
     {
-        assert_eq!(-9223372036854775808 as i64, EnumI64::VariantMin.to_primitive());
-        assert_eq!(9223372036854775807 as i64, EnumI64::VariantMax.to_primitive());
+        assert_eq!(-9223372036854775808_i64, EnumI64::VariantMin.to_primitive());
+        assert_eq!(9223372036854775807_i64, EnumI64::VariantMax.to_primitive());
     }
 }
